EquipmentTable and UserStorages
